chore: move TracingConfig init to tracing crate
